backend/requirements.txt
Django==3.2.20
djangorestframework==3.12.4
django-colorfield==0.10.1
django-cors-headers==3.13.0
django-filter==23.2
djoser==2.1.0
drf-extra-fields==3.6.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
Pillow==9.3.0
python-dotenv==1.0.0
urllib3==1.26.10